+++ HISTORY.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+2.0 - (2013-12-13)
+------------------
+
+* Feature: Entirely update password generator algorithm. Default 16 chars,
+  Google Application Specific Passwords like.
+* Feature: Possibility to import / export settings from / to CSV file.
+* Feature: Added more advanced options to control password generation.
+  Password length, Use or not numbers, special characters, capital letters.
+  Possibility to store and retrieve user name per service.
+  Possibility to change password with versions
+* Change: Use ilogin_old for previous generator/settings.
+
 0.2 - (2011-08-02)
 ------------------